refactor(skills): remove zeph-skills dependency on zeph-memory and zeph-tools (#2905)
Move shared domain types to zeph-common:
- SkillTrustLevel (from zeph-tools)
- ToolErrorCategory + ErrorDomain enums (from zeph-tools error_taxonomy)
- RAW_INJECTION_PATTERNS, strip_format_chars (from zeph-tools patterns)
- QUARANTINE_DENIED (from zeph-tools quarantine)

zeph-tools re-exports all moved types for backwards compatibility.
classify_* functions remain in zeph-tools.

Feature-gate Qdrant in zeph-skills:
- New optional feature: `qdrant = ["dep:zeph-memory", "dep:qdrant-client"]`
- SkillError::Qdrant, SkillMatcherBackend::Qdrant, and qdrant_matcher module
  are all #[cfg(feature = "qdrant")]
- zeph-core enables the qdrant feature explicitly

Default builds of zeph-skills now compile without zeph-memory or qdrant-client.

Closes #2905
