+++ crates/zeph-skills/src/qdrant_matcher.rs
@@ -1,6 +1,8 @@
 // SPDX-FileCopyrightText: 2026 Andrei G <bug-ops>
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
+#![cfg(feature = "qdrant")]
+
 pub use zeph_memory::SyncStats;
 use zeph_memory::{Embeddable, EmbeddingRegistry, QdrantOps};
